moved calculateHoursTutored to SessionService
