+++ services/SessionService.js
@@ -9,6 +9,7 @@ const { USER_BAN_REASON, SESSION_REPORT_REASON } = require('../constants')
 const UserActionCtrl = require('../controllers/UserActionCtrl')
 const ObjectId = require('mongodb').ObjectId
 const { USER_ACTION } = require('../constants')
+const VolunteerModel = require('../models/Volunteer')
 
 const addPastSession = async ({ userId, sessionId }) => {
   await User.update({ _id: userId }, { $addToSet: { pastSessions: sessionId } })
@@ -24,6 +25,58 @@ const isSessionParticipant = (session, user) => {
   return [session.student, session.volunteer].some(
     participant => !!participant && user._id.equals(participant)
   )
+}
+
+const calculateHoursTutored = session => {
+  const threeHoursMs = 1000 * 60 * 60 * 3
+  const fifteenMinsMs = 1000 * 60 * 15
+
+  const { volunteerJoinedAt, endedAt, messages, volunteer } = session
+  if (!volunteer) return 0
+  if (!(volunteerJoinedAt && endedAt)) return 0
+  // skip if no messages are sent
+  if (messages.length === 0) return 0
+
+  const volunteerJoinDate = new Date(volunteerJoinedAt)
+  const sessionEndDate = new Date(endedAt)
+  let sessionLengthMs = sessionEndDate - volunteerJoinDate
+
+  // skip if volunteer joined after the session ended
+  if (sessionLengthMs < 0) return 0
+
+  let latestMessageIndex = messages.length - 1
+  let wasMessageSentAfterSessionEnded =
+    messages[latestMessageIndex].createdAt > sessionEndDate
+
+  // get the latest message that was sent within a 15 minute window of the message prior.
+  // Sometimes sessions are not ended by either participant and one of the participants may send
+  // a message to see if the other participant is still active before ending the session.
+  // Exclude these messages when getting the total session end time
+  if (sessionLengthMs > threeHoursMs || wasMessageSentAfterSessionEnded) {
+    while (
+      latestMessageIndex > 0 &&
+      (wasMessageSentAfterSessionEnded ||
+        messages[latestMessageIndex].createdAt -
+          messages[latestMessageIndex - 1].createdAt >
+          fifteenMinsMs)
+    ) {
+      latestMessageIndex--
+      wasMessageSentAfterSessionEnded =
+        messages[latestMessageIndex].createdAt > sessionEndDate
+    }
+  }
+
+  const latestMessageDate = new Date(messages[latestMessageIndex].createdAt)
+
+  // skip if the latest message was sent before a volunteer joined
+  // or skip if the only messages that were sent were after a session has ended
+  if (latestMessageDate <= volunteerJoinDate || wasMessageSentAfterSessionEnded)
+    return 0
+
+  sessionLengthMs = latestMessageDate - volunteerJoinDate
+
+  // milliseconds in an hour = (60,000 * 60) = 3,600,000
+  return Number((sessionLengthMs / 3600000).toFixed(2))
 }
 
 module.exports = {
@@ -83,11 +136,13 @@ module.exports = {
       sessionId: session._id
     })
 
-    if (session.volunteer)
-      await addPastSession({
-        userId: session.volunteer,
-        sessionId: session._id
-      })
+    if (session.volunteer) {
+      const hoursTutored = calculateHoursTutored(session)
+      await VolunteerModel.updateOne(
+        { _id: session.volunteer },
+        { $addToSet: { pastSessions: session._id }, $inc: { hoursTutored } }
+      )
+    }
 
     const quillDoc = await QuillDocService.getDoc(session._id.toString())
     const whiteboardDoc = await WhiteboardService.getDoc(session._id.toString())
@@ -496,5 +551,7 @@ module.exports = {
     } catch (err) {
       throw new Error(err.message)
     }
-  }
+  },
+
+  calculateHoursTutored
 }
